third_party/blink/web_tests/external/wpt/css/css-shapes/shape-functions/circle-function-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['shape-outside'] = "circle(20px at center)" should set the property value
  assert_equals: serialization should be canonical expected "circle(20px at center center)" but got "circle(20px at 50% 50%)"
[FAIL] e.style['shape-outside'] = "circle(4% at top right)" should set the property value
  assert_equals: serialization should be canonical expected "circle(4% at right top)" but got "circle(4% at 100% 0%)"
[FAIL] e.style['shape-outside'] = "circle(closest-corner at center)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "circle(closest-corner at 20px 50px)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "circle(closest-side at center)" should set the property value
  assert_equals: serialization should be canonical expected "circle(at center center)" but got "circle(at 50% 50%)"
[FAIL] e.style['shape-outside'] = "circle(farthest-corner at center top)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "circle(farthest-corner at center)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['shape-outside'] = "circle(farthest-side at center top)" should set the property value
  assert_equals: serialization should be canonical expected "circle(farthest-side at center top)" but got "circle(farthest-side at 50% 0%)"
[FAIL] e.style['shape-outside'] = "circle(farthest-side at center)" should set the property value
  assert_equals: serialization should be canonical expected "circle(farthest-side at center center)" but got "circle(farthest-side at 50% 50%)"
Harness: the test ran to completion.
